Changed all so it works on all operating systems
Created script notifier.py that creates a notification window
using Customtkinter.

Because trying to make notifications in all systems natively was
complicated and some OS's don't give much flexibility like using
.gif and overall window structure.

The script also handles sound depending on the system.

Updated anatra.py so it detects a keypress in the loop of the stopwatch
natively.

Testing on Linux and MacOs still to be done.